requirements.txt
Django==5.1.3
djangorestframework==3.15.2
environs==11.0.0
dj-database-url==2.3.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
